+++ lib/src/modules/audio_output.rs
@@ -3,6 +3,8 @@ use core::cell::RefCell;
 
 use graphity::Node;
 
+use crate::core::signal::Signal;
+
 pub struct AudioOutput {
     buffer: Rc<RefCell<[f32; 32]>>,
 }
@@ -24,11 +26,11 @@ pub struct AudioOutputConsumer;
 #[derive(PartialEq, Eq, Hash, Clone, Copy, Debug)]
 pub enum AudioOutputProducer {}
 
-impl Node<[f32; 32]> for AudioOutput {
+impl Node<Signal> for AudioOutput {
     type Consumer = AudioOutputConsumer;
     type Producer = AudioOutputProducer;
 
-    fn write(&mut self, _consumer: Self::Consumer, input: [f32; 32]) {
-        *self.buffer.borrow_mut() = input;
+    fn write(&mut self, _consumer: Self::Consumer, input: Signal) {
+        *self.buffer.borrow_mut() = input.as_audio();
     }
 }